Initial commit of CircularLog support
This adds a basic CircularLog class. It has a log() method which can
be called like printf() to store formatted text output in a RAM based
circular log.  The user can later call the dump() method providing a
FILE* to which the last bit of the log should be dumped.  The oldest
entry will most likely be missing some text at its beginning.  The
clear() method can be used to reset the log.

The CircularLog class has template parameters.  The first is how many
bytes should be reserved for the circular log so that it can be
allocated at compile time.  The second is the maximum number of bytes
to be reserved on the stack when formatting the log output for a single
log() call.  If this is too small then longer lines will be truncated
before they are added to the circular log.

I added a makefile in the test/ directory which can build and run the
unit tests with an optional code coverage analysis.

I updated .gitignore to handle temporary unit test and code coverage
output.