Make CircularLog::isEmpty() public
This method already existed but previously it was protected. Making it
public allows the SDFileSystem (which uses a CircularLog as an error
log) to expose the current state of its error log to an application.
An application can then know if internal errors were detected within
the driver and display them to the user.