SPIDma driver for LPC17xx
The LPC17xx implementation of the SPIDma class which I wrote the
SDFileSystem to utilize. It takes advantage of the FIFO and DMA
functionality of the LPC17xx SSP peripheral to help improve
performance.

This commit also includes LoopbackTest which exercises the
various SPIDma methods. Before attempting to build and use this
test, a couple of things should be noted:
* You must set SPIDMA_LOOP_BACK_TEST to 1 in SPIDma/SPIDma.h as this
  enables a circular queue of the discarded reads so that they can be
  verified by the test. I wouldn't enable this for general use since
  it just uses up RAM and CPU cycles for no reason.
* The p5 and p6 pins of the mbed-LPC1768 should be shorted together
  to loop MOSI back into MISO. There is a flag to do this in the
  periperhal itself but doing it manually was more thorough and the
  chip select needed to be manually looped back anyway.
* The p7 and p8 pins of the mbed-LPC1768 should be shorted together
  so that the external state of the chip select pin can be verified
  by the test.