Fix sector count calculation for SDv2 cards
I was looking at the various test run results and noticed that the disk
size logged seemed like they were off by a factor of 1024 for SDv2
cards. It turns out that I didn't notice that C_SIZE was in units of
1024 blocks for SDv2 cards.

This only impacted information displayed to the user since the
FATFileSystem only asks for this information while formating a card.
During normal read/write operations, it is never used.

I updated the pre-built binaries as well.