Add FATFileSystem::fflush() method
There are situations, like logging, where the user wants to make sure
that all of the data they have written so far has made it to media so
that it won't be lost if the device crashes or loses power. There is an
expectation that calling the Standard C Library's fflush() routine
alone would be enough to accomplish this but unfortunately, it isn't.
From what I have seen, both GCC/Newlib and Keil/ARMlibC have an
implementation of fflush() which make sure that any data sitting in the
FILE*'s buffers get flushed down to the underlying file system but they
don't do anything to trigger the same file system to flush its
outstanding writes to non-volatile storage. At one time it appears that
ARM used the __sys_ensure() syscall for this but that has since been
deprecated and is no longer called. Newlib doesn't seem to make it easy
to accomplish either.

This commit is my attempt to work around this issue:
* First the user calls the Standard C Library's fflush() API to make
  sure that all outstanding writes have been flushed down to the
  FATFileSystem driver.
* Then the user can call the fflush() method on their FATFileSystem
  object and it will call the fsync() method on all of the opened
  FATFileHandles. To accomplish this I now maintain a linked list of
  active handles. Handles are inserted into the list at open()
  time and then removed from the list at close() time.