Remove counters no longer required
I updated the README to summarize the results of the PerformanceTest
runs so far.  Based on this summary, I decided to remove two of the
counters:
* selectFirstExchangeRequiredCount
* cmd12PaddingByteRequiredCount

I had added those counters to determine if certain code was required.
The test results have already shown that the code is required so I have
removed the related instrumentation.