Don't deselect() before CMD12 during read
Previously I was de-asserting the chip select pin before sending the
CMD12 command to stop a multi-block read. This meant that the
subsequent select() would waitWhileBusy() during the receipt of an
actual block's data. This appears to be what was causing the
selectFirstExchangeRequiredCount counter to be non-zero. Moving the
deselect() call to after the CMD12 command appears to have eliminated
the need for that dummy read in select(). I had originally written the
code like this because of how ChaN had done things in his LPC1768
demo.

This code change has actually resulted in a >10% performance
improvement for reads.

I also modified the cmd12PaddingByteRequiredCount counter to only
increment if the discarded response would have been detected as an
R1 error response. This shows that without this discarded response
during CMD12 processing, the read path would have encountered an
error. Since the test runs haven't shown any read failures, this code
must be necessary since it appears to have discarded errors that would
have otherwise resulted in errors being found in the test log.