Removed Timer usage
The SDFileSystem driver had a few places where it used mbed Timer
objects to place an upper limit on the time it should wait for the SD
card to respond. This made it difficult to debug the code via the MRI
debug monitor when the driver would end up timing out operations once
the debug monitor resumed execution since the Timer objects didn't
pause during debug monitor halts. It could also happen in a real
application if a higher priority ISR used the CPU for some time
while a SD operation was in progress.

This commit replaces those timers with maximum loop iteration counts.
These limits are scaled based on the current SPI frequency so that the
new limits will be similar to the old version when nothing interrupts
the code running within the loop.

This does mean that the maximumACMD41LoopTime(),
maximumWaitWhileBusyTime(), and maximumReceiveDataBlockWaitTime()
counters are now just estimates based on loop iterations.